refactor: create test folder, move tests in it
